+++ src/main/kotlin/cloud/glitchdev/rfu/config/categories/TrophyFishing.kt
@@ -9,6 +9,13 @@ object TrophyFishing : Category("Trophy Fishing") {
     override val description: TranslatableValue
         get() = Literal("Settings for Trophy Fishing!")
 
+    init {
+        dualSeparator {
+            title = "Pity"
+            description = "Features related to Pity"
+        }
+    }
+
     var trophyPityDisplay by reloadableBoolean(true) {
         name = Literal("Trophy Pity Display")
         description = Literal("Displays trophy pity progress on the screen.")
@@ -41,5 +48,17 @@ object TrophyFishing : Category("Trophy Fishing") {
     var appendPityToTrophyMessage by boolean(true) {
         name = Literal("Append Pity to Trophy Message")
         description = Literal("Appends the count before pity to the trophy catch message in chat (e.g. '(100)').")
+    }
+
+    init {
+        dualSeparator {
+            title = "Trophy Specific"
+            description = "Features related to specific trophies"
+        }
+    }
+
+    var slugfishTimer by boolean(false) {
+        name = Literal("Slugfish Timer")
+        description = Literal("Displays the 20s timer required to catch slugfish above the bobber, shorter if using slug pet.")
     }
 }