+++ src/main/kotlin/cloud/glitchdev/rfu/gui/hud/elements/PetDisplay.kt
@@ -17,9 +17,7 @@ object PetDisplay : AbstractTextHudElement("petDisplay") {
     private var currentPet : String? = null
 
     override fun onInitialize() {
-        currentPet = PetUpdateEventManager.currentPet
-
-        registerPetUpdateEvent { pet ->
+        registerPetUpdateEvent { pet, _, _ ->
             currentPet = pet
             updateState()
         }